requirements.txt
flask
gunicorn
requests
discord.py
python-dotenv
